v3.29.2 - Combined elements into GameItems - Seperated Decision (easier imports)
