+++ TeamCode/src/main/java/display/Display.java
@@ -11,7 +11,7 @@ import math.calculus.Differentiator;
 import math.calculus.Integrator;
 
 public class Display extends Drawer {
-    // TODO FIX DISPLAY
+    // TODO 4 FIX DISPLAY
 
     private static final int width = 700;
     private static final int height = 700;
