--- TeamCode/src/main/java/elements/Ball.java
@@ -1,14 +1,0 @@
-package elements;
-
-
-public class Ball {
-    /**
-     * Radius and diameter of the balls
-     */
-    public static final double radius = 3.5; //cm
-    public static final double diameter = 2*radius;
-    /**
-     * Weight of a ball
-     */
-    public static final double weight = 33.5; //g
-}
