--- TeamCode/src/main/java/elements/Barrier.java
@@ -1,11 +1,0 @@
-package elements;
-
-/**
- * NOTE: Uncommented
- */
-
-public class Barrier {
-    // TOD4 FIX ARNAV
-    // Add data about barrier here
-    // Length radius, geometric representation
-}
