--- TeamCode/src/main/java/elements/Box.java
@@ -1,14 +1,0 @@
-package elements;
-
-public class Box {
-    /**
-     * Width of the cube (or height since cube)
-     */
-    public static final double width = 5.1; //cm
-    /**
-     * Weights of different cubes
-     */
-    public static final double lightWeight = 50.5; //g
-    public static final double mediumWeight = 93.1; //g
-    public static final double heavyWeight = 135.4; //g
-}
