--- TeamCode/src/main/java/elements/Field.java
@@ -1,12 +1,0 @@
-package elements;
-
-import geometry.CoordinatePlane;
-
-/**
- * NOTE: Uncommented
- */
-
-public class Field {
-    // TOD4 NEW VARINI
-    // Have geometric represntation of everything in here
-}
