+++ TeamCode/src/main/java/geometry/GeometryObject.java
@@ -8,7 +8,7 @@ import static java.lang.Math.*;
  * NOTE: Uncommented
  */
 
-// TODO FIX GEOMETRY PACKAGE
+// TODO 4 FIX GEOMETRY PACKAGE
 
 public abstract class GeometryObject {
     public GeometryObject getRelativeTo(Pose origin) {
