+++ TeamCode/src/main/java/util/condition/Decision.java
@@ -1,0 +1,6 @@
+package util.condition;
+
+/**
+ * Implemented if the class represents a type of decision
+ */
+public interface Decision {}
